HBAI-227 Move and rename all embeddings funcs
